+++ apps/meteor/client/views/admin/subscription/components/UpgradeButton.tsx
@@ -2,16 +2,18 @@ import { Button, ButtonGroup, Throbber } from '@rocket.chat/fuselage';
 import type { ButtonProps } from '@rocket.chat/fuselage/dist/components/Button/Button';
 import type { ReactElement } from 'react';
 import React, { memo } from 'react';
-import { useTranslation } from 'react-i18next';
 
 import { useCheckoutUrlAction } from '../hooks/useCheckoutUrl';
 
-type UpgradeButtonProps = {
-	i18nKey?: string;
-} & Partial<ButtonProps>;
-
-const UpgradeButton = ({ i18nKey = 'Manage_subscription', ...props }: UpgradeButtonProps): ReactElement => {
-	const { t } = useTranslation();
+const UpgradeButton = ({
+	children,
+	target = '_blank',
+	action,
+	...props
+}: Partial<ButtonProps> & {
+	target: string;
+	action: string;
+}): ReactElement => {
 	const mutation = useCheckoutUrlAction();
 
 	const handleBtnClick = () => {
@@ -19,13 +21,16 @@ const UpgradeButton = ({ i18nKey = 'Manage_subscription', ...props }: UpgradeBut
 			return;
 		}
 
-		mutation.mutate();
+		mutation.mutate({
+			target,
+			action,
+		});
 	};
 
 	return (
 		<ButtonGroup align='end'>
 			<Button onClick={() => handleBtnClick()} {...props} disabled={mutation.isLoading}>
-				{mutation.isLoading ? <Throbber inheritColor size='x12' /> : t(i18nKey)}
+				{mutation.isLoading ? <Throbber inheritColor size='x12' /> : children}
 			</Button>
 		</ButtonGroup>
 	);