+++ apps/meteor/client/views/admin/subscription/components/cards/MACCard.tsx
@@ -1,11 +1,22 @@
+import { Palette } from '@rocket.chat/fuselage';
 import type { ReactElement } from 'react';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
 import type { CardProps } from '../FeatureUsageCard';
-import PieGraphCard from '../PieGraphCard';
+import FeatureUsageCard from '../FeatureUsageCard';
+import UpgradeButton from '../UpgradeButton';
+import UsagePieGraph from '../UsagePieGraph';
 
-const MACCard = ({ value = 0, max }: { value: number; max: number }): ReactElement => {
+const MACCard = ({
+	value = 0,
+	max,
+	hideManageSubscription,
+}: {
+	value: number;
+	max: number;
+	hideManageSubscription?: boolean;
+}): ReactElement => {
 	const { t } = useTranslation();
 
 	const pieGraph = {
@@ -15,14 +26,37 @@ const MACCard = ({ value = 0, max }: { value: number; max: number }): ReactEleme
 
 	const nearLimit = pieGraph && pieGraph.used / pieGraph.total >= 0.8;
 
+	const macLeft = pieGraph.total - pieGraph.used;
+
 	const card: CardProps = {
 		title: t('Monthly_active_contacts'),
 		infoText: t('MAC_InfoText'),
-		showUpgradeButton: nearLimit,
-		upgradeButtonText: 'Buy_more',
+		...(hideManageSubscription && {
+			upgradeButton: (
+				<UpgradeButton target='mac-card' action='buy_more' small>
+					{t('Buy_more')}
+				</UpgradeButton>
+			),
+
+			...(nearLimit && {
+				upgradeButton: (
+					<UpgradeButton target='mac-card' action='upgrade' small>
+						{t('Upgrade')}
+					</UpgradeButton>
+				),
+			}),
+		}),
 	};
 
-	return <PieGraphCard pieGraph={pieGraph} card={card} />;
+	const color = nearLimit ? Palette.statusColor['status-font-on-danger'].toString() : undefined;
+
+	const message = macLeft > 0 ? t('MAC_Available', { macLeft }) : undefined;
+
+	return (
+		<FeatureUsageCard card={card}>
+			<UsagePieGraph label={message} used={pieGraph.used} total={pieGraph.total} color={color} />
+		</FeatureUsageCard>
+	);
 };
 
 export default MACCard;