+++ apps/meteor/ee/server/services/CHANGELOG.md
@@ -1,5 +1,174 @@
 # rocketchat-services
 
+## 1.1.15-rc.7
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.7
+- @rocket.chat/rest-typings@6.5.0-rc.7
+- @rocket.chat/core-services@0.3.0-rc.7
+- @rocket.chat/model-typings@0.2.0-rc.7
+- @rocket.chat/models@0.0.21-rc.7
+
+## 1.1.15-rc.6
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.6
+- @rocket.chat/rest-typings@6.5.0-rc.6
+- @rocket.chat/core-services@0.3.0-rc.6
+- @rocket.chat/model-typings@0.2.0-rc.6
+- @rocket.chat/models@0.0.21-rc.6
+
+## 1.1.15-rc.5
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.5
+- @rocket.chat/rest-typings@6.5.0-rc.5
+- @rocket.chat/core-services@0.3.0-rc.5
+- @rocket.chat/model-typings@0.2.0-rc.5
+- @rocket.chat/models@0.0.21-rc.5
+
+## 1.1.15-rc.4
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.4
+- @rocket.chat/rest-typings@6.5.0-rc.4
+- @rocket.chat/core-services@0.3.0-rc.4
+- @rocket.chat/model-typings@0.2.0-rc.4
+- @rocket.chat/models@0.0.21-rc.4
+
+## 1.1.15-rc.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.3
+- @rocket.chat/rest-typings@6.5.0-rc.3
+- @rocket.chat/core-services@0.3.0-rc.3
+- @rocket.chat/model-typings@0.2.0-rc.3
+- @rocket.chat/models@0.0.21-rc.3
+
+## 1.1.15-rc.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.2
+- @rocket.chat/rest-typings@6.5.0-rc.2
+- @rocket.chat/core-services@0.3.0-rc.2
+- @rocket.chat/model-typings@0.2.0-rc.2
+- @rocket.chat/models@0.0.21-rc.2
+
+## 1.1.15-rc.1
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.1
+- @rocket.chat/rest-typings@6.5.0-rc.1
+- @rocket.chat/core-services@0.3.0-rc.1
+- @rocket.chat/model-typings@0.2.0-rc.1
+- @rocket.chat/models@0.0.21-rc.1
+
+## 1.1.15-rc.0
+
+### Patch Changes
+
+- Updated dependencies [dea1fe9191]
+- Updated dependencies [c0ef13a0bf]
+- Updated dependencies [223dce18a3]
+- Updated dependencies [5b9d6883bf]
+- Updated dependencies [92613680b7]
+- Updated dependencies [ec1b2b9846]
+- Updated dependencies [a98f3ff303]
+- Updated dependencies [5f81a0f3cb]
+- Updated dependencies [dea1fe9191]
+  - @rocket.chat/core-typings@6.5.0-rc.0
+  - @rocket.chat/model-typings@0.2.0-rc.0
+  - @rocket.chat/core-services@0.3.0-rc.0
+  - @rocket.chat/rest-typings@6.5.0-rc.0
+  - @rocket.chat/models@0.0.21-rc.0
+
+## 1.1.17
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.8
+- @rocket.chat/rest-typings@6.4.8
+- @rocket.chat/core-services@0.2.8
+- @rocket.chat/model-typings@0.1.8
+- @rocket.chat/models@0.0.23
+
+## 1.1.16
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.7
+- @rocket.chat/rest-typings@6.4.7
+- @rocket.chat/core-services@0.2.7
+- @rocket.chat/model-typings@0.1.7
+- @rocket.chat/models@0.0.22
+
+## 1.1.15
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.6
+- @rocket.chat/rest-typings@6.4.6
+- @rocket.chat/core-services@0.2.6
+- @rocket.chat/model-typings@0.1.6
+- @rocket.chat/models@0.0.21
+
+## 1.1.14
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.5
+- @rocket.chat/rest-typings@6.4.5
+- @rocket.chat/core-services@0.2.5
+- @rocket.chat/model-typings@0.1.5
+- @rocket.chat/models@0.0.20
+
+## 1.1.13
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.4
+- @rocket.chat/rest-typings@6.4.4
+- @rocket.chat/core-services@0.2.4
+- @rocket.chat/model-typings@0.1.4
+- @rocket.chat/models@0.0.19
+
+## 1.1.12
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.3
+- @rocket.chat/rest-typings@6.4.3
+- @rocket.chat/core-services@0.2.3
+- @rocket.chat/model-typings@0.1.3
+- @rocket.chat/models@0.0.18
+
+## 1.1.11
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.2
+- @rocket.chat/rest-typings@6.4.2
+- @rocket.chat/core-services@0.2.2
+- @rocket.chat/model-typings@0.1.2
+- @rocket.chat/models@0.0.17
+
+## 1.1.10
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.1
+- @rocket.chat/rest-typings@6.4.1
+- @rocket.chat/core-services@0.2.1
+- @rocket.chat/model-typings@0.1.1
+- @rocket.chat/models@0.0.16
+
 ## 1.1.9
 
 ### Patch Changes